Move digest functionality to modelator::util::digest (#41)
